Use size_t when makes sense

STC simplification:
ELO   | 1.33 +- 2.90 (95%)
SPRT  | 10.0+0.10s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [-3.00, 1.00]
GAMES | N: 28800 W: 7623 L: 7513 D: 13664

Bench: 3091908

+++ src/core/board_state.h
@@ -27,6 +27,6 @@ namespace core {
         Zobrist hash = Zobrist();
         Piece piece_captured = NULL_PIECE;
         CastlingRights rights = CastlingRights();
-        unsigned int move50 = 0;
+        size_t move50 = 0;
     };
 } // namespace core
